Add unit testing

+++ database/migrations/2023_06_06_152351_create_product_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->float('price');
+            $table->double('price', 15, 2);
             $table->string('description');
             $table->string('image');
             $table->boolean('in_stock');
